move progress bar to separate folder
