+++ client/src/pages/OnBoarding/SetAvailability.js
@@ -11,7 +11,7 @@ import {
   makeStyles,
 } from "@material-ui/core";
 
-import ProgressBar from "./ProgressBar";
+import ProgressBar from "./smallComponents/ProgressBar";
 import AvailableHours from "./smallComponents/AvailableHoursButton";
 
 const SetAvailability = (props) => {
