+++ client/src/pages/OnBoarding/SetTimezoneUrl.js
@@ -9,7 +9,7 @@ import {
   makeStyles,
 } from "@material-ui/core";
 
-import ProgressBar from "./ProgressBar";
+import ProgressBar from "./smallComponents/ProgressBar";
 
 const SetTimezoneUrl = (props) => {
   //styles
