+++ client/src/pages/OnBoarding/smallComponents/ProgressBar.js
